# Performance improvements for Options