--- a/Tests/Common/Data/Market/QuoteBarTests.cs
+++ b/Tests/Common/Data/Market/QuoteBarTests.cs
@@ -1,11 +1,11 @@
 ﻿/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -47,30 +47,6 @@ namespace QuantConnect.Tests.Common.Data.Market
             Assert.AreEqual(11, bar.High);
             Assert.AreEqual(11, bar.Low);
             Assert.AreEqual(11, bar.Close);
-        }
-
-        [Test]
-        public void QuoteBarReader_CanParseTradeBarFormattedData_Successfully()
-        {
-            var config = new SubscriptionDataConfig(typeof(TradeBar), Symbols.SPY, Resolution.Second, TimeZones.NewYork, TimeZones.NewYork, false, false, false);
-            var line = "14280000,1.10905,1.10909,1.10904,1.10908";
-            var date = DateTime.MaxValue;
-            var isLiveMode = false;
-
-            var quoteBar = new QuoteBar();
-            var parsedQuoteBar = (QuoteBar) quoteBar.Reader(config, line, date, isLiveMode);
-
-            Assert.AreEqual(parsedQuoteBar.Symbol, Symbols.SPY);
-
-            Assert.AreEqual(parsedQuoteBar.Ask.Open, 1.10905);
-            Assert.AreEqual(parsedQuoteBar.Ask.High, 1.10909);
-            Assert.AreEqual(parsedQuoteBar.Ask.Low, 1.10904);
-            Assert.AreEqual(parsedQuoteBar.Ask.Close, 1.10908);
-
-            Assert.AreEqual(parsedQuoteBar.Bid.Open, 1.10905);
-            Assert.AreEqual(parsedQuoteBar.Bid.High, 1.10909);
-            Assert.AreEqual(parsedQuoteBar.Bid.Low, 1.10904);
-            Assert.AreEqual(parsedQuoteBar.Bid.Close, 1.10908);
         }
 
         [Test]