<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- `QuoteBar.cs`: Replace `Split` for simple for loop, avoids creating unnecessary substrings.
- `SecurityIdentifier` will calculate `StrikePrice` just once. Replacing
`!=` for a direct call to `Equals()` preventing unnecessary checks.
- Slightly improving Linq queries at `OptionFilterUniverse`
- Replace `OpenReader().CopyTo` for `Extract(stream)`, avoids copying the
data twice.
- Replace `DateTime.Now` for `DateTime.UtcNow`

**Performance tests**
```
        public override void Initialize()
        {
            SetStartDate(2018, 1, 1);
            SetEndDate(2018, 6, 1);
            SetCash(10000);

            var option = AddOption("GOOG");

            option.SetFilter(-8, +8, TimeSpan.FromDays(20), TimeSpan.FromDays(50));
        }
```
`master`
243.22 seconds at 553k data points per second. Processing total of 134,502,627 data points.
233.58 seconds at 576k data points per second. Processing total of 134,502,627 data points.
225.68 seconds at 596k data points per second. Processing total of 134,502,627 data points.

`PR ~30% improvement`
154.34 seconds at 871k data points per second. Processing total of 134,502,627 data points.
149.04 seconds at 902k data points per second. Processing total of 134,502,627 data points.
147.59 seconds at 911k data points per second. Processing total of 134,502,627 data points.

```
        public override void Initialize()
        {
            SetStartDate(2018, 1, 1);
            SetEndDate(2018, 2, 1);
            SetCash(10000);

            var option = AddOption("GOOG");

            option.SetFilter(
                universe => universe.IncludeWeeklys().Strikes(-8, +8).Expiration(TimeSpan.FromDays(20), TimeSpan.FromDays(50))
            );
        }
```
`master`
105.30 seconds at 293k data points per second. Processing total of 30,883,838 data points.
92.11 seconds at 335k data points per second. Processing total of 30,883,838 data points.
100.15 seconds at 308k data points per second. Processing total of 30,883,838 data points.
`PR ~25% improvements`
75.10 seconds at 411k data points per second. Processing total of 30,883,838 data points.
75.32 seconds at 410k data points per second. Processing total of 30,883,838 data points.
75.73 seconds at 408k data points per second. Processing total of 30,883,838 data points.
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Related to https://github.com/QuantConnect/Lean/issues/3283
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- Faster algorithms trading Options
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->